Move activities to .activities package
