+++ app/src/main/AndroidManifest.xml
@@ -10,15 +10,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".GameList" />
-        <activity android:name=".CreateGame"></activity>
+        <activity android:name=".activities.GameListActivity" />
+        <activity android:name=".activities.CreateGameActivity"></activity>
     </application>
 
 </manifest>
