+++ app/src/main/java/ch/epfl/sweng/erpa/activities/MainActivity.java
@@ -1,9 +1,13 @@
-package ch.epfl.sweng.erpa;
+package ch.epfl.sweng.erpa.activities;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
+
+import ch.epfl.sweng.erpa.R;
+import ch.epfl.sweng.erpa.activities.CreateGameActivity;
+import ch.epfl.sweng.erpa.activities.GameListActivity;
 
 public class MainActivity extends Activity {
 
@@ -15,13 +19,13 @@ public class MainActivity extends Activity {
 
     public void launch_game_list(View view)
     {
-        Intent intent = new Intent(this, GameList.class);
+        Intent intent = new Intent(this, GameListActivity.class);
         startActivity(intent);
     }
 
     public void launch_create_game(View view)
     {
-        Intent intent = new Intent(this, CreateGame.class);
+        Intent intent = new Intent(this, CreateGameActivity.class);
         startActivity(intent);
     }
 }
